chore: Move Tournament tests to new directory
